graphql documents re-organised
