+++ litmus-portal/frontend/src/graphql/mutations.ts
@@ -1,0 +1,47 @@
+import { gql } from '@apollo/client';
+
+export const CREATE_WORKFLOW = gql`
+  mutation createChaosWorkFlow($ChaosWorkFlowInput: ChaosWorkFlowInput!) {
+    createChaosWorkFlow(input: $ChaosWorkFlowInput) {
+      workflow_id
+      cronSyntax
+      workflow_name
+      workflow_description
+      isCustomWorkflow
+    }
+  }
+`;
+export const CREATE_USER = gql`
+  mutation CreateUser($user: UserInput!) {
+    createUser(user: $user) {
+      id
+      username
+      created_at
+      updated_at
+      removed_at
+    }
+  }
+`;
+
+export const SEND_INVITE = gql`
+  mutation sendInvite($member: MemberInput!) {
+    sendInvitation(member: $member) {
+      user_id
+      user_name
+      role
+      invitation
+    }
+  }
+`;
+
+export const ACCEPT_INVITE = gql`
+  mutation accept($member: MemberInput!) {
+    acceptInvitation(member: $member)
+  }
+`;
+
+export const DECLINE_INVITE = gql`
+  mutation decline($member: MemberInput!) {
+    declineInvitation(member: $member)
+  }
+`;
