+++ litmus-portal/frontend/src/graphql/quries.ts
@@ -1,0 +1,65 @@
+import { gql } from '@apollo/client';
+
+export const WORKFLOW_DETAILS = gql`
+  query workflowDetails($projectID: String!) {
+    getWorkFlowRuns(project_id: $projectID) {
+      workflow_id
+      workflow_name
+      workflow_run_id
+      execution_data
+      project_id
+      cluster_name
+      last_updated
+    }
+  }
+`;
+
+export const GET_USER = gql`
+  query getUser($username: String!) {
+    getUser(username: $username) {
+      username
+      email
+      name
+      projects {
+        members {
+          user_id
+          user_name
+          role
+          invitation
+          name
+          email
+          joined_at
+        }
+        name
+        id
+      }
+      company_name
+      updated_at
+      created_at
+      removed_at
+      is_email_verified
+      state
+      role
+    }
+  }
+`;
+
+export const GET_CLUSTER = gql`
+  query getCluster($project_id: String!, $cluster_type: String) {
+    getCluster(project_id: $project_id, cluster_type: $cluster_type) {
+      cluster_id
+      is_active
+    }
+  }
+`;
+
+export const ALL_USERS = gql`
+  query allUsers {
+    users {
+      id
+      name
+      username
+      email
+    }
+  }
+`;
