+++ litmus-portal/frontend/src/graphql/subscriptions.ts
@@ -1,0 +1,16 @@
+/* eslint-disable import/prefer-default-export */
+import { gql } from '@apollo/client';
+
+export const WORKFLOW_EVENTS = gql`
+  subscription workflowEvents($projectID: String!) {
+    workflowEventListener(project_id: $projectID) {
+      workflow_id
+      workflow_name
+      workflow_run_id
+      execution_data
+      project_id
+      cluster_name
+      last_updated
+    }
+  }
+`;
